% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{named}
\alias{named}
\title{Named
This is just an example of a function with named slots.
Run the function with named arguments and the values
of the inputs will be printed to the console}
\usage{
named(arg1 = "default input 1", arg2 = "default input 2",
  arg3 = "default input 3")
}
\description{
Named
This is just an example of a function with named slots.
Run the function with named arguments and the values
of the inputs will be printed to the console
}
